# packages/civic-protocol-core/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Civic Ledger API
  description: |
    The Civic Ledger API provides endpoints for managing Reflections, Attestations, and Votes
    in the Civic Protocol ecosystem. This API enables civic participation through verified
    activity and governance mechanisms.
  version: 0.1.0
  contact:
    name: Civic Protocol Core
    url: https://github.com/civic-protocol/core
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5411
    description: Development server
  - url: https://api.civicprotocol.org
    description: Production server
paths:
...
  /balance/{address}:
    get:
      summary: Get MIC balance
      description: Retrieve the MIC balance for a given address
      parameters:
        - name: address
          in: path
          required: true
          description: Citizen or companion address
          schema:
            type: string
      responses:
        '200':
          description: Balance information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /earn/events:
    get:
      summary: Get earn events
      description: Retrieve MIC earning events for an address
      parameters:
        - name: address
          in: query
          required: true
          description: Citizen or companion address
          schema:
            type: string
        - name: date
          in: query
          description: Filter by date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: limit
          in: query
          description: Maximum number of events to return
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of earn events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/EarnEvent'
                  total:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
...
    Balance:
      type: object
      properties:
        address:
          type: string
          description: Citizen or companion address
        balance:
          type: string
          description: MIC balance (as string to handle large numbers)
        vesting:
          type: string
          description: Vesting balance
        nonce:
          type: integer
          description: Transaction nonce
        last_updated:
          type: string
          format: date-time
          description: Last balance update
    EarnEvent:
      type: object
      properties:
        event_id:
          type: string
          description: Unique event identifier
        address:
          type: string
          description: Recipient address
        amount:
          type: string
          description: MIC amount earned
        reason:
          type: string
          description: Reason for earning
        cycle_id:
          type: string
          description: Associated cycle ID
        created_at:
          type: string
          format: date-time
          description: Event timestamp
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "BAD_REQUEST"
            message: "Invalid request parameters"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "NOT_FOUND"
            message: "Resource not found"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "INTERNAL_ERROR"
            message: "An internal error occurred"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []